pyproject.toml
[tool.poetry]
name = "playhub"
version = "0.1.0"
description = ""
authors = ["6691a <6691a@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
gunicorn = "^21.2.0"
uvicorn = {extras = ["standard"], version = "^0.28.1"}
pyyaml = "^6.0.1"
sentry-sdk = "^1.42.0"
django-structlog = "^8.0.0"
django-simple-history = "^3.5.0"
django-storages = {extras = ["google"], version = "^1.14.3"}
django-allauth = "^0.61.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
psycopg2-binary = "^2.9.9"
django-redis = "^5.4.0"
confluent-kafka = "^2.3.0"
werkzeug = "^3.0.1"
babel = "^2.15.0"
pydantic = "^2.7.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
django-stubs = "^4.2.7"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
black = "^24.3.0"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__,", "*/migrations/*"]
ignore = ["F403", "F405"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
plugins = [
    "mypy_django_plugin.main",
]
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
exclude = ["^migrations$", "^settings.py$", "^manage.py$"]

[tool.django-stubs]
django_settings_module = "config.settings.settings"

[tool.isort]
profile = "black"
